Run CI on Ruby 2.2, 2.3, 2.4, 2.5 and JRuby 9.1 on CircleCI 2.0
This switches us from using Travis CI to CircleCI, which is our
standard choice across GoCardless.

As before, we run our tests on Ruby 2.3, 2.3, 2.4, 2.5 and JRuby
9.1.

This *stops* running tests against [Rubinius][1] (known as `rbx`
in Travis). It wasn't passing anyway, and it seems that current
versions of Rubocop are incompatible with it anyway, as Rubocop
now depends on [ruby-progressbar][2] which [uses][3] Ruby 2.x's
[refinements][4] feature which is [not supported][5] by
Rubinius.

To get there, we:

* Start excluding `vendor/` from Rubocop, since we'll install
gems to there in Circle
* Add `rspec_junit_formatter` to the gemspec as a development
dependency to get nicely-formatted results in Circle
* Include a `.circleci/config.yml` file to configure CircleCI
* Delete the `.travis.yml` file, so we'll no longer run tests
in Travis

[1]: https://rubinius.com/
[2]: https://github.com/jfelchner/ruby-progressbar
[3]: https://github.com/jfelchner/ruby-progressbar/blob/master/lib/ruby-progressbar/refinements/enumerator.rb#L4
[4]: https://docs.ruby-lang.org/en/2.4.0/syntax/refinements_rdoc.html
[5]: http://www.rubydoc.info/github/rubinius/rubinius