Add support for coercing to a date
Currently, if you want to pass a date in an env var, you need to
interpret it as a string and parse that string yourself.  This means
you need to handle validation failures at runtime, which is less than
ideal.  We recently hit this in payments-service.

By adding a date type here, we can validate date env vars in the same
way as bool and int vars.