Add support for loading a datetime value
Currently, there is no way to pass a datetime environment variable. The closest thing we have is the `:date` type which one can only use for dates. We have recently hit a use case where it would be ideal to support loading `:datetime` environment variables (we want to set a date and time after which some business logic changes).

This commit adds just that + the validation required to load a datetime value correctly.